# docs/swagger/fpabil-1.0.0-swagger.yaml
swagger: '2.0'
info:
  title: FPABIL
  description: Gestione abilitazione iride
  version: 1.0.0
  contact:
    email: andrea1.fontanelli@consulenti.csi.it
# host: tst-applogic.reteunitaria.piemonte.it
# basePath: /fpabilbl/rest/1.0.0
...
paths:
...
  /anagrafiche/comuni/byprovincia:
    get:
      summary: Get comuni By provicia 
      description: Restituisce l'elenco dei comuni della
                    provincia passata come
      operationId: getComuniByProvincia
      produces:
      - application/json
      - application/xml
      parameters:
        - in: query
          name: codprovincia
          type: string
          required: true
      responses:
        200:
          schema: 
            $ref: '#/definitions/ComuneArray'
          description: OK   
...
host: virtserver.swaggerhub.com
...
basePath: /andreafontanelli/fpabilbl/1.0.0
...
definitions:
...
  Provincia:
    type: object
    description: L'oggetto rappresentante la provincia
    properties:
      codice:
        type: string
        description: codice provincia
      descrizione:
        type: string
        description: la descrizione della provincia
      sigla:
        type: string
        description: la sigla della provincia
...
  Comune:
    type: object
    description: L'oggetto rappresentante il comune
    properties:
      codice:
        type: string
        description: codice comune
      descrizione:
        type: string
        description: la descrizione del comune
      provincia:
        $ref: '#/definitions/Provincia'
        description: la provincia di appartenenza al Comune  
  ComuneArray:
    type: array
    items:
      $ref: '#/definitions/Comune'
    description: Elenco di oggetti comune
...
schemes:
 - https
 - http